@@ -1,10 +1,7 @@
-"use strict";
+import checksum from "checksum";
+import BaseMigration from "./BaseMigration.js";
 
-const checksum = require("checksum");
-
-const BaseMigration = require("./BaseMigration");
-
-module.exports = class CqlMigration extends BaseMigration {
+export default class CqlMigration extends BaseMigration {
 
   constructor(opts) {
     super("CQL", opts);
@@ -16,13 +13,14 @@ module.exports = class CqlMigration extends BaseMigration {
   }
 
   async execute(cassandraClients) {
-    let cqlStatements = this.cql.split(/;[\r\n]+/);
-    for (let cs of cqlStatements)
+    let cqlStatements = this.cql.trim().split(/;[\r\n]+/);
+    for (let cs of cqlStatements) {
       await this.executeStatement(cassandraClients[0], cs);
+    }
   }
 
   async executeStatement(cassandraClient, cs) {
     await cassandraClient.execute(cs);
   }
 
-};
+}
