@@ -1,18 +1,15 @@
+import checksum from "checksum";
+import BaseMigration from "./BaseMigration.js";
 
-"use strict";
-
-const checksum = require("checksum");
-
-const BaseMigration = require("./BaseMigration");
-
-module.exports = class JsMigration extends BaseMigration {
+export default class JsMigration extends BaseMigration {
 
   constructor(opts) {
     super("JS", opts);
   }
 
   async load() {
-    this.fn = require(this.opts.path);
+    const { default: module } = await import(this.opts.path);
+    this.fn = module;
     this.src = await BaseMigration.loadFile(this.opts.path);
     this.checksum = checksum(this.src);
   }
@@ -21,4 +18,4 @@ module.exports = class JsMigration extends BaseMigration {
     return this.fn(cassandraClients);
   }
 
-};
+}
