@@ -1,16 +1,14 @@
-"use strict";
+import debug from "debug";
+import { readdir } from "fs/promises";
+import { sep } from "path";
+import EventEmitter from "events";
+import JsMigration from "./JsMigration.js";
+import CqlMigration from "./CqlMigration.js";
 
-const debug = require("debug")("cassandra-shift");
-
-const fs = require("fs");
-const path = require("path");
-const EventEmitter = require("events");
-
-const JsMigration = require("./JsMigration");
-const CqlMigration = require("./CqlMigration");
+const csDebug = debug("cassandra-shift");
 
 async function ensureKeyspace(cassandraClient, keyspace) {
-  debug(`Ensuring keyspace ${keyspace}, if not exists it will be created with default settings`);
+  csDebug(`Ensuring keyspace ${keyspace}, if not exists it will be created with default settings`);
 
   await createDefaultKeyspace(cassandraClient, keyspace);
 }
@@ -28,13 +26,13 @@ async function dropKeyspace(cassandraClient, keyspace) {
 }
 
 async function useKeyspace(cassandraClient, keyspace) {
-  debug(`Using keyspace ${keyspace}`);
+  csDebug(`Using keyspace ${keyspace}`);
 
   await cassandraClient.execute(`USE ${keyspace}`);
 }
 
 async function ensureMigrationTable(cassandraClient, migrationTable) {
-  debug(`Ensuring table ${migrationTable} exists`);
+  csDebug(`Ensuring table ${migrationTable} exists`);
 
   await cassandraClient.execute(`CREATE TABLE IF NOT EXISTS ${migrationTable} (
     version int,
@@ -49,7 +47,7 @@ async function ensureMigrationTable(cassandraClient, migrationTable) {
 }
 
 async function getAppliedMigrations(cassandraClient, migrationTable) {
-  debug("Retrieving table applied migration list");
+  csDebug("Retrieving table applied migration list");
 
   return (await cassandraClient.execute(`SELECT * FROM ${migrationTable}`)).rows.sort(function(r1, r2) {
     return r1.version - r2.version;
@@ -70,12 +68,12 @@ async function loadAvailableMigration(dir, filename) {
   let type = match[3];
     
   let opts = {
-    path: dir + path.sep + filename,
+    path: dir + sep + filename,
     version: parseInt(match[1]),
     name: prettyName(match[2])
   };
 
-  debug(`Loading migration definition ${filename}`);
+  csDebug(`Loading migration definition ${filename}`);
 
   let migration;
   switch(type) {
@@ -93,9 +91,9 @@ async function loadAvailableMigration(dir, filename) {
 }
 
 async function loadAvailableMigrations(dir) {
-  debug("Loading available migrations");
+  csDebug("Loading available migrations");
 
-  let migrationFiles = fs.readdirSync(dir);
+  let migrationFiles = await readdir(dir);
 
   let migrations = [];
   for (let file of migrationFiles) {
@@ -128,7 +126,7 @@ function computeState(appliedMigration, migration) {
 }
 
 function checkAppliedMigrations(appliedMigrations, migrations) {
-  debug("Checking applied migrations");
+  csDebug("Checking applied migrations");
 
   if (appliedMigrations.length === 0)
     return;
@@ -138,7 +136,7 @@ function checkAppliedMigrations(appliedMigrations, migrations) {
     throw new Error(`Migration ${lm.version} - "${lm.name}" failed, fix manually before retrying`);
 
   if (appliedMigrations.length > migrations.length) {
-    debug(`Applied ${appliedMigrations.length} migrations, but only ${migrations.length} defined`);
+    csDebug(`Applied ${appliedMigrations.length} migrations, but only ${migrations.length} defined`);
   }
 
   for (let i = 0; i < appliedMigrations.length; ++i) {
@@ -160,7 +158,7 @@ function checkAppliedMigrations(appliedMigrations, migrations) {
 }
 
 async function executeMigration(cassandraClients, migrationTable, migration) {
-  debug(`Applying migration ${migration.version} "${migration.name}"`);
+  csDebug(`Applying migration ${migration.version} "${migration.name}"`);
 
   let now = new Date();
 
@@ -193,7 +191,7 @@ async function executeMigration(cassandraClients, migrationTable, migration) {
     throw cachedError;
 }
 
-module.exports = class Shift extends EventEmitter {
+export default class Shift extends EventEmitter {
 
   constructor(cassandraClients, opts = {}) {
     super();
@@ -234,17 +232,17 @@ module.exports = class Shift extends EventEmitter {
       });
     }
 
-    debug("Migrations applied succeffully");
+    csDebug("Migrations applied succeffully");
   }
 
   async clean() {
-    debug(`Cleaning keyspace ${this.opts.keyspace}"`);
+    csDebug(`Cleaning keyspace ${this.opts.keyspace}"`);
 
     await dropKeyspace(this.cassandraClients[0], this.opts.keyspace);
     await createDefaultKeyspace(this.cassandraClients[0], this.opts.keyspace);
     this.emit("cleaned");
 
-    debug("Keyspace cleaned succeffully");
+    csDebug("Keyspace cleaned succeffully");
   }
 
   async info() {
@@ -308,4 +306,4 @@ module.exports = class Shift extends EventEmitter {
       return false;
     }
   }
-};
+}
